\name{Data_potato}
\alias{Data_potato}
\docType{data}
\title{
Normalized Monthly Average Potato Price of India
}
\description{
Normalized Monthly Average Potato Price of India from January 2010 to July 2020.
}
\usage{data("Data_potato")}
\format{
  A time series data with 127 observations.
  \describe{
    \item{\code{price}}{a time series}
  }
}
\details{
Dataset contains 127 observations of normalized monthly average potato price of India. It is obtained from World Bank "Pink sheet".
}
\source{
Department of Consumer Affairs, Govt. of India
}
\references{
https://consumeraffairs.nic.in/
}
\examples{
data(Data_potato)
}
\keyword{datasets}
